WIP: debug LDFs

+++ src/interfaces/ILiquidityDensityFunction.sol
@@ -74,7 +74,7 @@ interface ILiquidityDensityFunction {
         bool useTwap,
         bytes32 ldfParams,
         bytes32 ldfState
-    ) external view returns (bool success, int24 roundedTick, uint256 cumulativeAmount, uint128 swapLiquidity);
+    ) external view returns (bool success, int24 roundedTick, uint256 cumulativeAmount, uint256 swapLiquidity);
 
     /// @notice Checks if the given LDF parameters are valid.
     /// @param tickSpacing The tick spacing of the pool
