+++ src/ldf/CarpetedGeometricDistribution.sol
@@ -64,7 +64,7 @@ contract CarpetedGeometricDistribution is ILiquidityDensityFunction {
         external
         pure
         override
-        returns (bool success, int24 roundedTick, uint256 cumulativeAmount, uint128 swapLiquidity)
+        returns (bool success, int24 roundedTick, uint256 cumulativeAmount, uint256 swapLiquidity)
     {
         (int24 minTick, int24 length, uint256 alphaX96, uint256 weightMain, ShiftMode shiftMode) =
             LibCarpetedGeometricDistribution.decodeParams(twapTick, key.tickSpacing, useTwap, ldfParams);
